+++ go/teams/seitan_v2_test.go
@@ -413,3 +413,85 @@ func TestTeamHandleMultipleSeitans(t *testing.T) {
 		require.Equal(t, keybase1.TeamInviteCategory_SEITAN, invtype)
 	}
 }
+
+func TestTeamInviteSeitanV2Failures(t *testing.T) {
+	tc := SetupTest(t, "team", 1)
+	defer tc.Cleanup()
+
+	tc.Tp.SkipSendingSystemChatMessages = true
+
+	user2, err := kbtest.CreateAndSignupFakeUser("team", tc.G)
+	require.NoError(t, err)
+	kbtest.Logout(tc)
+
+	admin, err := kbtest.CreateAndSignupFakeUser("team", tc.G)
+	require.NoError(t, err)
+
+	teamName, teamID := createTeam2(tc)
+
+	t.Logf("Created team %q", teamName.String())
+
+	token, err := CreateSeitanTokenV2(context.Background(), tc.G,
+		teamName.String(), keybase1.TeamRole_WRITER, keybase1.SeitanKeyLabel{})
+	require.NoError(t, err)
+
+	t.Logf("Created token %q", token)
+
+	kbtest.LogoutAndLoginAs(tc, user2)
+
+	// Generate invitation id, but make Signature with different IKey.
+	// Simulate "replay attack" or similar.
+	ikey, err := ParseIKeyV2FromString(string(token))
+	require.NoError(t, err)
+	sikey, err := ikey.GenerateSIKey()
+	require.NoError(t, err)
+	inviteID, err := sikey.GenerateTeamInviteID()
+	require.NoError(t, err)
+
+	ikey2, err := GenerateIKeyV2() // ikey2 is not the ikey from token.
+	require.NoError(t, err)
+	sikey2, err := ikey2.GenerateSIKey()
+	require.NoError(t, err)
+	now := keybase1.ToTime(time.Now())
+	_, maliciousPayload, err := sikey2.GenerateSignature(user2.GetUID(), user2.EldestSeqno, inviteID, now)
+	require.NoError(t, err)
+
+	err = postSeitanV2(tc.MetaContext(), acceptedSeitanV2{
+		encoded:  maliciousPayload,
+		now:      now,
+		inviteID: SCTeamInviteID(inviteID),
+	})
+	require.NoError(t, err)
+
+	teamInviteID, err := inviteID.TeamInviteID()
+	require.NoError(t, err)
+
+	t.Logf("handle synthesized rekeyd command")
+	kbtest.LogoutAndLoginAs(tc, admin)
+
+	msg := keybase1.TeamSeitanMsg{
+		TeamID: teamID,
+		Seitans: []keybase1.TeamSeitanRequest{{
+			InviteID:    teamInviteID,
+			Uid:         user2.GetUID(),
+			EldestSeqno: user2.EldestSeqno,
+			Akey:        keybase1.SeitanAKey(maliciousPayload),
+			Role:        keybase1.TeamRole_WRITER,
+			UnixCTime:   int64(now),
+		}},
+	}
+	err = HandleTeamSeitan(context.Background(), tc.G, msg)
+	// Seitan handler does not fail, but ignores the request.
+	require.NoError(t, err)
+
+	t.Logf("invite should still be there")
+	t0, err := GetTeamByNameForTest(context.Background(), tc.G, teamName.String(), false /* public */, true /* needAdmin */)
+	require.NoError(t, err)
+	require.Equal(t, 1, t0.NumActiveInvites(), "invite should still be active")
+	require.EqualValues(t, t0.CurrentSeqno(), 2)
+
+	t.Logf("user should not be in team")
+	role, err := t0.MemberRole(context.Background(), user2.GetUserVersion())
+	require.NoError(t, err)
+	require.Equal(t, keybase1.TeamRole_NONE, role, "user role")
+}
